fix(@angular/cli): fix problem of SuppressPlugin in case entryFiles type is string (#7393)
The type of entryFiles from entryPoint is array in webpack official suggested, but actually it could be string. The error caused by `every` method which is only invoked by array, however entryFiles is a string. Solution: if entryFiles is not array, make it as an array which includes just a single string element.