fix(@ngtools/webpack): allow path mapping to resolve in JS files
Previously we filtered our path mapping resolver to only be used on typescript.
The correct behaviour is unclear; tsc does not resolve JS files, only TS. But
there is a lot of value to use path mapping to resolve JavaScript files, and
it replaces the webpack alias configuration option. Because of that value it
was decided to fix this.

Fixes #8117.